+++ src/deep-flat/solution/sandbox.js
@@ -1,9 +1,9 @@
 //import { solutionName } from './solution-name.js';
 export const solutionName = (nestedArray = []) => {
-    return nestedArray.flat(Infinity);
-  };
+  return nestedArray.flat(Infinity);
+};
 
-  describe('Flatten a nested array', () => {
+describe('Flatten a nested array', () => {
   it('[1, 2, [3]] => [1, 2, 3]', () => {
     expect(solutionName([1, 2, [3]])).toEqual([1, 2, 3]);
   });
