--- src/split-object/SolutionDotEntries/Boludez.js
@@ -1,10 +1,0 @@
-const slipknot = {
-    Album1999: 'Slipknot',
-    Album2001: 'Iowa',
-    };
-    
-    console.log(slipknot);
-    
-    const arr = Object.entries(slipknot);
-
-    console.log(arr);
