--- src/split-object/SolutionDotEntries/solutionDotEntries.js
@@ -1,6 +1,0 @@
-/**
- * Splitting an Object into and Array of Arrays.
- * @param {object} 
- * @returns {Array} Array of arrays. 
- */
-export const solutionName = () => {};
