+++ src/split-object/solution-dot-entries/Boludez.js
@@ -1,0 +1,10 @@
+const slipknot = {
+  Album1999: 'Slipknot',
+  Album2001: 'Iowa',
+};
+
+console.log(slipknot);
+
+const arr = Object.entries(slipknot);
+
+console.log(arr);
