+++ src/split-object/solution-dot-entries/README.md
@@ -1,7 +1,6 @@
-# Split Object: Object.entries() Method  
+# Split Object: Object.entries() Method
 
 Take an object and split it into an array of objects. Each entry in the array will be one key/value pair from the object.
-
 
 <!-- BEGIN DOCS -->
 
@@ -9,9 +8,9 @@ Take an object and split it into an array of objects. Each entry in the array wi
 
 ---
 
-## Strategy  
+## Strategy
 
-At first glance it seems logical to just use the `.Split()` Method, but the challenge is specific with the result. So in this case it is going to be more useful to implement the built-in 
+At first glance it seems logical to just use the `.Split()` Method, but the challenge is specific with the result. So in this case it is going to be more useful to implement the built-in
 
 ---
 
@@ -19,11 +18,11 @@ At first glance it seems logical to just use the `.Split()` Method, but the chal
 
 ---
 
-## Use Cases  
+## Use Cases
 
 In the video of reference there is a mention of working with JSON packages in which you have to import or extract data from Objects.
 
-1. Maybe a most common use case could be to access data like scores from a particular mission in a video game. 
+1. Maybe a most common use case could be to access data like scores from a particular mission in a video game.
 2. Access data like ingredients and quantities for a recipe.
 
 ---
@@ -32,7 +31,6 @@ In the video of reference there is a mention of working with JSON packages in wh
 
 This video was really useful for understanding the built-in method and possible **Use Cases**
 [Link to the video](https://www.youtube.com/watch?v=A0g7sga28Zg)
-
 
 <!--
   was there any code, blog post, video, ... that inspired your solution?
