+++ src/split-object/solution-dot-entries/sandbox.js
@@ -1,0 +1,15 @@
+import { solutionName } from "./solution-dot-entries.js";
+
+/*
+import { solutionName } from './solutionDotEntries.js';
+
+export const solutionName = (solutionDotEntries = []) => {
+    return solutionDotEntries.entries();
+};
+*/
+const slipknot = {
+Album1999: 'Slipknot',
+Album2001: 'Iowa',
+};
+
+console.log(Object.entries(slipknot));
