+++ src/split-object/solution-dot-entries/solution-dot-entries.js
@@ -1,0 +1,11 @@
+/**
+ * Splitting an Object into and Array of Arrays.
+ * @param {object} [toSeparate={}]
+ * @returns {Array} Array of arrays.
+ * @example
+ * splitObject({ a: 1, b: 2, c: 3 });
+ * // -> [{ a: 1 }, { b: 2 }, { c: 3 }]
+ */
+export const objectSplitter = Object.entries();
+
+// Object.entries
